feat(global): support set lang in query
